reorganized, edited action space
organized my file space to be neater, found out that position control will change the direct angles of the motor joints, which is what we want since it's what we can control.  Removed joint velocity from state space since we can't get it directly in physical robot, so we won't use it, although technically I think I can derive it.