reoriented robot to more manageable start position
Calibrating the physical robot motor positions is still a giant pain since I don't know a good way to do it, so I'm initializing the robot in a very angular way to make it easier on me - more parallel and perpendicular arrangements for the legs. Also did some reorganizing and updated the URDF to match the new config, will need to retrain on this model now. First, need to update the leg ranges again.